Add back oauth and ldap auth provider.
