+++ dhis-2/dhis-services/dhis-service-core/src/main/resources/META-INF/dhis/security.xml
@@ -105,6 +105,11 @@
   <!-- Security : Authentication providers -->
 
   <sec:authentication-manager alias="authenticationManager">
+    <sec:authentication-provider user-service-ref="userDetailsService">
+      <sec:password-encoder ref="bCryptPasswordEncoder" />
+    </sec:authentication-provider>
     <sec:authentication-provider ref="twoFactorAuthenticationProvider" />
+    <sec:authentication-provider ref="ldapAuthenticationProvider" />
+    <sec:authentication-provider user-service-ref="clientDetailsUserService" />
   </sec:authentication-manager>
 </beans>
