+++ dhis-2/dhis-web/dhis-web-commons/src/main/resources/META-INF/dhis/security.xml
@@ -440,5 +440,5 @@
 
   <bean id="appsSystemAuthoritiesProvider" class="org.hisp.dhis.security.authority.AppsSystemAuthoritiesProvider" />
 
-  <bean id="authenticationListener" class="org.hisp.dhis.security.AuthenticationListener2FA"/>
+  <bean id="authenticationListener" class="org.hisp.dhis.security.AuthenticationListener"/>
 </beans>
